backend/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.3.0
bidict==0.23.1
cffi==2.0.0
click==8.2.1
cryptography==45.0.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
python-engineio==4.12.2
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.10.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.53
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
websockets==12.0
wsproto==1.2.0
